Fix structure by move app.ts and server.ts and moveout __test__ directory
