+++ package.json
@@ -24,7 +24,7 @@
     "ts-node": "^10.9.2"
   },
   "scripts": {
-    "dev": "nodemon server.ts",
+    "dev": "nodemon src/server.ts",
     "start": "ts-node server.ts",
     "test": "jest --detectOpenHandles --coverage",
     "test:coverage": "nyc jest --coverage",
